Unit testing of refactoring complete

+++ src/au/edu/wehi/idsv/EvidenceClusterProcessor.java
@@ -53,7 +53,7 @@ public class EvidenceClusterProcessor implements Iterable<VariantContextDirected
 				lowDir = highDir;
 				highDir = tmpDir;
 			}
-			GraphNode node = new GraphNode(startX, endX, startY, endY, PhredLogLikelihoodRatioModel.llr(evidence));
+			GraphNode node = new GraphNode(startX, endX, startY, endY, (float)PhredLogLikelihoodRatioModel.llr(evidence));
 			addNode(lowDir, highDir, node);
 		} else {
 			if (filterOut(loc)) return;
